dirs types'
