+++ python/ry/dirs.pyi
@@ -1,38 +1,38 @@
-def audio() -> str | None: ...
-def audio_dir() -> str | None: ...
-def cache() -> str | None: ...
-def cache_dir() -> str | None: ...
-def config() -> str | None: ...
-def config_dir() -> str | None: ...
-def config_local() -> str | None: ...
-def config_local_dir() -> str | None: ...
-def data() -> str | None: ...
-def data_dir() -> str | None: ...
-def data_local() -> str | None: ...
-def data_local_dir() -> str | None: ...
-def desktop() -> str | None: ...
-def desktop_dir() -> str | None: ...
-def document() -> str | None: ...
-def document_dir() -> str | None: ...
-def download() -> str | None: ...
-def download_dir() -> str | None: ...
-def executable() -> str | None: ...
-def executable_dir() -> str | None: ...
-def font() -> str | None: ...
-def font_dir() -> str | None: ...
-def home() -> str | None: ...
-def home_dir() -> str | None: ...
-def picture() -> str | None: ...
-def picture_dir() -> str | None: ...
-def preference() -> str | None: ...
-def preference_dir() -> str | None: ...
-def public() -> str | None: ...
-def public_dir() -> str | None: ...
-def runtime() -> str | None: ...
-def runtime_dir() -> str | None: ...
-def state() -> str | None: ...
-def state_dir() -> str | None: ...
-def template() -> str | None: ...
-def template_dir() -> str | None: ...
-def video() -> str | None: ...
-def video_dir() -> str | None: ...
+from ry.ryo3.dirs import audio as audio
+from ry.ryo3.dirs import audio_dir as audio_dir
+from ry.ryo3.dirs import cache as cache
+from ry.ryo3.dirs import cache_dir as cache_dir
+from ry.ryo3.dirs import config as config
+from ry.ryo3.dirs import config_dir as config_dir
+from ry.ryo3.dirs import config_local as config_local
+from ry.ryo3.dirs import config_local_dir as config_local_dir
+from ry.ryo3.dirs import data as data
+from ry.ryo3.dirs import data_dir as data_dir
+from ry.ryo3.dirs import data_local as data_local
+from ry.ryo3.dirs import data_local_dir as data_local_dir
+from ry.ryo3.dirs import desktop as desktop
+from ry.ryo3.dirs import desktop_dir as desktop_dir
+from ry.ryo3.dirs import document as document
+from ry.ryo3.dirs import document_dir as document_dir
+from ry.ryo3.dirs import download as download
+from ry.ryo3.dirs import download_dir as download_dir
+from ry.ryo3.dirs import executable as executable
+from ry.ryo3.dirs import executable_dir as executable_dir
+from ry.ryo3.dirs import font as font
+from ry.ryo3.dirs import font_dir as font_dir
+from ry.ryo3.dirs import home as home
+from ry.ryo3.dirs import home_dir as home_dir
+from ry.ryo3.dirs import picture as picture
+from ry.ryo3.dirs import picture_dir as picture_dir
+from ry.ryo3.dirs import preference as preference
+from ry.ryo3.dirs import preference_dir as preference_dir
+from ry.ryo3.dirs import public as public
+from ry.ryo3.dirs import public_dir as public_dir
+from ry.ryo3.dirs import runtime as runtime
+from ry.ryo3.dirs import runtime_dir as runtime_dir
+from ry.ryo3.dirs import state as state
+from ry.ryo3.dirs import state_dir as state_dir
+from ry.ryo3.dirs import template as template
+from ry.ryo3.dirs import template_dir as template_dir
+from ry.ryo3.dirs import video as video
+from ry.ryo3.dirs import video_dir as video_dir
