+++ python/ry/ryo3/dirs.pyi
@@ -1,0 +1,38 @@
+def audio() -> str | None: ...
+def audio_dir() -> str | None: ...
+def cache() -> str | None: ...
+def cache_dir() -> str | None: ...
+def config() -> str | None: ...
+def config_dir() -> str | None: ...
+def config_local() -> str | None: ...
+def config_local_dir() -> str | None: ...
+def data() -> str | None: ...
+def data_dir() -> str | None: ...
+def data_local() -> str | None: ...
+def data_local_dir() -> str | None: ...
+def desktop() -> str | None: ...
+def desktop_dir() -> str | None: ...
+def document() -> str | None: ...
+def document_dir() -> str | None: ...
+def download() -> str | None: ...
+def download_dir() -> str | None: ...
+def executable() -> str | None: ...
+def executable_dir() -> str | None: ...
+def font() -> str | None: ...
+def font_dir() -> str | None: ...
+def home() -> str | None: ...
+def home_dir() -> str | None: ...
+def picture() -> str | None: ...
+def picture_dir() -> str | None: ...
+def preference() -> str | None: ...
+def preference_dir() -> str | None: ...
+def public() -> str | None: ...
+def public_dir() -> str | None: ...
+def runtime() -> str | None: ...
+def runtime_dir() -> str | None: ...
+def state() -> str | None: ...
+def state_dir() -> str | None: ...
+def template() -> str | None: ...
+def template_dir() -> str | None: ...
+def video() -> str | None: ...
+def video_dir() -> str | None: ...
